Split up setup.py and move packaging tools to new script
