+++ .github/workflows/test_linux.yml
@@ -32,7 +32,7 @@ jobs:
         run: |
           pip install -U -r requirements.txt -r requirements-dev.txt
       - name: Run Build Commands
-        run: python setup.py qtlrelease sample
+        run: python pkgutils.py qtlrelease sample
       - name: Run Tests
         run: |
           export QT_QPA_PLATFORM=offscreen
