+++ i18n/README.md
@@ -76,7 +76,7 @@ The `i18n` folder at the root of the repository contains the `nw_XX.ts` translat
 Whether to add a new language to the translation framework, or to  update the file against the
 current source code, you must first run the `qtlupdate` command:
 ```bash
-python3 setup.py qtlupdate i18n/nw_XX.ts
+python3 pkgutils.py qtlupdate i18n/nw_XX.ts
 ```
 
 The file name must be in the correct format. All translation files must be located in the `i18n`
@@ -101,7 +101,7 @@ The application does not use `.ts` files directly. The `novelwriter/assets/i18n/
 the actual files used to translate the GUI into another language other than the default British
 English. These files are not generated by default, but they can be built with:
 ```bash
-python3 setup.py qtlrelease
+python3 pkgutils.py qtlrelease
 ```
 
 You can now test the translation in novelWriter. The Preferences dialog should list the newly added
