+++ setup/make_clean.sh
@@ -1,10 +1,10 @@
 #!/bin/bash
 set -e
 
-if [ ! -f setup.py ]; then
+if [ ! -f pkgutils.py ]; then
     echo "Must be called from the root folder of the source"
     exit 1
 fi
 
 rm -rfv dist_upload
-python3 setup.py build-clean
+python3 pkgutils.py build-clean
