+++ setup/make_release.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 set -e
 
-if [ ! -f setup.py ]; then
+if [ ! -f pkgutils.py ]; then
     echo "Must be called from the root folder of the source"
     exit 1
 fi
@@ -10,20 +10,20 @@ echo ""
 echo " Building Dependencies"
 echo "================================================================================"
 echo ""
-python3 setup.py clean-assets
-python3 setup.py qtlrelease manual sample
+python3 pkgutils.py clean-assets
+python3 pkgutils.py qtlrelease manual sample
 
 echo ""
 echo " Building Minimal Packages"
 echo "================================================================================"
 echo ""
-python3 setup.py minimal-zip --target-win
-python3 setup.py minimal-zip --target-linux
-python3 setup.py minimal-zip --target-darwin
+python3 pkgutils.py minimal-zip --target-win
+python3 pkgutils.py minimal-zip --target-linux
+python3 pkgutils.py minimal-zip --target-darwin
 
 echo ""
 echo " Building Linux Packages"
 echo "================================================================================"
 echo ""
-python3 setup.py build-deb --sign
-python3 setup.py build-ubuntu --sign --first
+python3 pkgutils.py build-deb --sign
+python3 pkgutils.py build-ubuntu --sign --first
