+++ setup/make_snapshot.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 set -e
 
-if [ ! -f setup.py ]; then
+if [ ! -f pkgutils.py ]; then
     echo "Must be called from the root folder of the source"
     exit 1
 fi
@@ -10,11 +10,11 @@ echo ""
 echo " Building Dependencies"
 echo "================================================================================"
 echo ""
-python3 setup.py clean-assets
-python3 setup.py qtlrelease manual sample
+python3 pkgutils.py clean-assets
+python3 pkgutils.py qtlrelease manual sample
 
 echo ""
 echo " Building Linux Snapshots"
 echo "================================================================================"
 echo ""
-python3 setup.py build-ubuntu --sign --snapshot
+python3 pkgutils.py build-ubuntu --sign --snapshot
